perf(webapp): componetization barChart
