+++ webapp/src/components/BarChartReport/index.js
@@ -1,8 +1,11 @@
-import React, { memo, useCallback, useState, useEffect } from 'react'
+import React, { memo, useCallback, useState } from 'react'
+import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
+import TooltipDownload from '@mui/material/Tooltip'
+import { useCurrentPng } from 'recharts-to-png'
+import { useTranslation } from 'react-i18next'
+import { makeStyles } from '@mui/styles'
 import PropTypes from 'prop-types'
 import FileSaver from 'file-saver'
-import { Box } from '@mui/system'
-import { makeStyles } from '@mui/styles'
 import {
   IconButton,
   Typography,
@@ -10,102 +13,11 @@ import {
   Switch,
   FormGroup
 } from '@mui/material'
-import TooltipDownload from '@mui/material/Tooltip'
-import DownloadOutlined from '@mui/icons-material/DownloadOutlined'
-import { useTranslation } from 'react-i18next'
-import { useCurrentPng } from 'recharts-to-png'
-import {
-  ComposedChart,
-  Bar,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-  ResponsiveContainer,
-  Cell
-} from 'recharts'
 
-import { formatWithThousandSeparator } from '../../utils/format-with-thousand-separator'
-
+import CustomBarchart from './custom-barchart'
 import styles from './styles'
 
 const useStyles = makeStyles(styles)
-
-const RenderChartLegend = ({ data }) => {
-  const classes = useStyles()
-  const { t } = useTranslation()
-
-  return (
-    <div className={classes.chartLegent}>
-      <a key={`key-Un${data.toLocaleLowerCase()}-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#f4d35e" />
-        {t(`un${data.toLocaleLowerCase()}`, { ns: 'generalForm' })}
-      </a>
-      <a key={`key-${data}-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#ee964b" />
-        {t(`${data.toLocaleLowerCase()}`, { ns: 'generalForm' })}
-      </a>
-      <a key={`key-total-link-chart`}>
-        <Box className={classes.legentCircle} bgcolor="#19647e" />
-        {t('total', { ns: 'generalForm' })}
-      </a>
-    </div>
-  )
-}
-
-RenderChartLegend.propTypes = {
-  data: PropTypes.string
-}
-
-const CustomTooltip = ({ payload = [], label = '', coinType = '' }) => {
-  const { t } = useTranslation()
-  label = label + ''
-  const arrayLabel = label.split(' ')
-
-  return (
-    <div>
-      <strong>
-        {`${t(arrayLabel[0].toLocaleLowerCase(), { ns: 'generalForm' })} ${
-          arrayLabel[1]
-        }`}
-      </strong>
-      {payload &&
-        payload.map((data, i) => (
-          <div key={`${i}-tooltip`}>
-            <div>
-              {i === 0 &&
-                data.payload?.date &&
-                `${
-                  data.payload.date ? t('date', { ns: 'generalForm' }) : ''
-                }: ${
-                  data.payload.date ? data.payload.date.split('T')[0] : ''
-                } `}
-            </div>
-            <div>
-              {`${t(
-                data.payload.category
-                  ? `${data.dataKey
-                      .split('_')[1]
-                      .toLocaleLowerCase()}${data.payload.category.toLocaleLowerCase()}`
-                  : data.dataKey.split('_')[1].toLocaleLowerCase(),
-                { ns: 'generalForm' }
-              )}: ${formatWithThousandSeparator(
-                data.payload[data.dataKey],
-                4
-              )} ${coinType}`}
-            </div>
-          </div>
-        ))}
-    </div>
-  )
-}
-
-CustomTooltip.propTypes = {
-  payload: PropTypes.array,
-  label: PropTypes.any,
-  coinType: PropTypes.string
-}
 
 const BarChartReport = ({
   data,
@@ -116,11 +28,8 @@ const BarChartReport = ({
 }) => {
   const classes = useStyles()
   const [getBarPng, { ref: barRef }] = useCurrentPng()
-  const [category, setCategory] = useState('')
   const { t } = useTranslation()
   const [selectedUSD, setSelected] = useState(false)
-  const [coinType, setCoinType] = useState('EOS')
-  const width = window.innerWidth
 
   const handleChange = event => {
     setSelected(event.target.checked)
@@ -133,15 +42,6 @@ const BarChartReport = ({
       FileSaver.saveAs(png, 'bar-chart.png')
     }
   }, [getBarPng])
-
-  useEffect(() => {
-    if (typeData === 'income') setCategory('Claimed')
-    else setCategory('Categorized')
-  }, [typeData])
-
-  useEffect(() => {
-    selectedUSD ? setCoinType('USD') : setCoinType('EOS')
-  }, [selectedUSD])
 
   return (
     <div className={classes.root}>
@@ -167,59 +67,12 @@ const BarChartReport = ({
         </div>
       </div>
       <div className={classes.chartContainer}>
-        <ResponsiveContainer width="100%" height={300} marginTop="16px">
-          <ComposedChart height={300} data={data} ref={barRef}>
-            <CartesianGrid stroke="#f0f0f0" />
-            <XAxis
-              tick={{ fontSize: 10, stroke: '#000', strokeWidth: 0.5 }}
-              dataKey="election"
-              scale="auto"
-            />
-            <YAxis
-              tick={{ fontSize: '10px', stroke: '#000', strokeWidth: 0.1 }}
-            />
-            <Tooltip
-              wrapperStyle={{
-                outline: 'none',
-                borderRadius: '4px',
-                backgroundColor: '#F9F9F9',
-                fontSize: '14px',
-                padding: '8px'
-              }}
-              content={<CustomTooltip coinType={coinType} />}
-            />
-            {showLegend && (
-              <Legend content={<RenderChartLegend data={category} />} />
-            )}
-            <Bar
-              dataKey={`${coinType}_UN${category.toLocaleUpperCase()}`}
-              barSize={width > 600 ? 35 : 15}
-              fill="#f4d35e"
-            >
-              {data.map(({ election }) => (
-                <Cell key={`cell-${election}`} />
-              ))}
-            </Bar>
-            <Bar
-              dataKey={`${coinType}_${category.toLocaleUpperCase()}`}
-              barSize={width > 600 ? 35 : 15}
-              fill="#ee964b"
-            >
-              {data.map(({ election }) => (
-                <Cell key={`cell-${election}`} />
-              ))}
-            </Bar>
-            <Bar
-              dataKey={`${coinType}_TOTAL`}
-              barSize={width > 600 ? 35 : 15}
-              fill="#19647e"
-            >
-              {data.map(({ election }) => (
-                <Cell key={`cell-${election}`} />
-              ))}
-            </Bar>
-          </ComposedChart>
-        </ResponsiveContainer>
+        <CustomBarchart
+          barRef={barRef}
+          showLegend={showLegend}
+          data={data}
+          typeData={typeData}
+        />
       </div>
     </div>
   )
